requirements.txt
fastapi
uvicorn
scrapling[all]
sentence-transformers
faiss-cpu
groq
python-multipart
curl_cffi